[internal] Rename hook files for naming consistency
