+++ src/client/components/Menu/UserLinks.tsx
@@ -13,7 +13,7 @@ import Link from "next/link";
 
 import styles from "@styles/Menu/navbar.module.scss";
 import { IUserPublic } from "@typing/user.interface";
-import { useLogout } from "@hooks/useLogout.hook";
+import { useLogout } from "@hooks/logout.hook";
 import { NavDropdown } from "react-bootstrap";
 
 interface Props {
