+++ src/client/hooks/logout.hook.ts
