+++ src/pages/user/index.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { useLoginRedirect } from "@hooks/useLoginRedirect.hook";
+import { useLoginRedirect } from "@hooks/login-redirect.hook";
 
 const index: FunctionComponent = () => {
   const user = useLoginRedirect();
